requirements.txt
django==5.2.6
django-silk==5.4.3

pillow==11.3.0

python-decouple==3.8

psycopg-binary==3.2.10

celery==5.4.0
celery-singleton==0.3.1
flower==2.0.1

djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
dj-rest-auth==7.0.1
django-allauth==65.11.2

drf-nested-routers==0.95.0

drf-spectacular==0.28.0

deepl==1.22.0

cryptography==46.0.1

django-taggit==6.1.0

black==25.9.0